roboclaw:Fix types

+++ src/drivers/roboclaw/RoboClaw.cpp
@@ -491,7 +491,7 @@ int RoboClaw::_transaction(e_command cmd, uint8_t *wbuff, size_t wbytes,
 	int count = write(_uart, buf, wbytes);
 
 	if (count < (int) wbytes) { // Did not successfully send all bytes.
-		PX4_ERR("Only wrote %d out of %d bytes", count, (int) wbytes);
+		PX4_ERR("Only wrote %d out of %zu bytes", count, wbytes);
 		return -1;
 	}
 
@@ -568,7 +568,7 @@ void RoboClaw::_parameters_update()
 		orb_set_interval(_actuatorsSub, _parameters.actuator_write_period_ms);
 	}
 
-	int baudRate;
+	long baudRate;
 	param_get(_param_handles.serial_baud_rate, &baudRate);
 
 	switch (baudRate) {
